Refactor : 파일 구조 변경 -> domain 위주
