+++ src/main/java/org/prgrms/vouchermanagement/blackCustomer/domain/BlackCustomer.java
@@ -1,4 +1,4 @@
-package org.prgrms.vouchermanagement.blackCustomer;
+package org.prgrms.vouchermanagement.blackCustomer.domain;
 
 import java.util.UUID;
