--- src/main/java/org/prgrms/vouchermanagement/customer/CustomerRepository.java
@@ -1,9 +1,0 @@
-package org.prgrms.vouchermanagement.customer;
-
-import java.util.List;
-
-public interface CustomerRepository {
-
-    List<Customer> showCustomerList();
-
-}
