+++ src/main/java/org/prgrms/vouchermanagement/customer/repository/CustomerRepository.java
@@ -1,0 +1,11 @@
+package org.prgrms.vouchermanagement.customer.repository;
+
+import org.prgrms.vouchermanagement.customer.domain.Customer;
+
+import java.util.List;
+
+public interface CustomerRepository {
+
+    List<Customer> showCustomerList();
+
+}
