+++ src/main/java/org/prgrms/vouchermanagement/voucher/policy/FixedAmountVoucher.java
@@ -1,4 +1,4 @@
-package org.prgrms.vouchermanagement.voucher;
+package org.prgrms.vouchermanagement.voucher.policy;
 
 public class FixedAmountVoucher implements DiscountPolicy {
