+++ src/main/java/org/prgrms/vouchermanagement/voucher/repository/MemoryRepository.java
@@ -1,6 +1,7 @@
-package org.prgrms.vouchermanagement.repository;
+package org.prgrms.vouchermanagement.voucher.repository;
 
-import org.prgrms.vouchermanagement.voucher.*;
+import org.prgrms.vouchermanagement.voucher.domain.Voucher;
+import org.prgrms.vouchermanagement.voucher.policy.DiscountPolicy;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
